requirements.txt
# requirements.txt

fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
pydantic==1.10.11
google-generativeai==0.1.0
pinecone-client==2.2.0
pymongo==4.4.0
pytest==7.4.0
pdfplumber==0.7.6
python-docx==0.8.11
python-multipart==0.0.6
